pyproject.toml
[project]
name = "realtime-ml-inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "pandas",
    "scikit-learn",
    "skl2onnx",
    "onnx",
    "packaging>=25.0",
    "sklearn2pmml",
    "pyyaml",
]
